requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.15.0
jinja2>=3.1.2
scikit-learn>=1.3.0